Move expr and stmt lowering into their own TUs
Cuts lower_expr.cpp (lower_expr_impl, eval_int, shape_of,
extrema_storage and the view/shape helpers) and lower_stmt.cpp
(lower_stmt_impl, needs_runtime_control, runtime_loop_control,
lower_runtime_ifelse, lower_runtime_int_sum, peel_terminal_return, the
island compiler) out of lower.cpp, which now holds only the
constructor, bind_data, lower_read_param, and the run/run_passes
pipeline. lower_structured_loop.inc stays included mid-class in
lower_internal.hpp, unchanged.

ctest: 120/120 passed.
dump_ops op-graph identity: 12/12 identical.
prep-trace census: diff -r census_before census_w2c_c3 -> IDENTICAL.
