+++ CMakeLists.txt
@@ -304,6 +304,8 @@ set(STANLI_RUNTIME_SOURCES
   runtime/src/mir_decode.cpp
   runtime/src/portable_mir_v2_reader.cpp
   runtime/src/lower.cpp
+  runtime/src/lower_expr.cpp
+  runtime/src/lower_stmt.cpp
   runtime/src/lower_funapp.cpp
   runtime/src/lower_higher_order.cpp
   runtime/src/structured_loop.cpp
